+++ RELEASE_NOTES.md
@@ -16,7 +16,9 @@ The library and demo app still updates way to often, and updates too much. For i
 
 ### ✨ New Features
 
+* `RichTextDataFormat` has a new `fileFormatText` property.
 * `RichTextDataFormat` has a new `libraryFormats` property that returns all formats except the `vendorArchivedData` format.
+* `RichTextDataFormatMenu` is a new menu for exporting rich text with various formats.
 * `RichTextDataReader` now has ways to get data for the current rich text.
 * `ViewDebug` is a new demo app class that is used for view debugging capabilities.
 * The library is now localized in Danish, German and Norwegian, although some translations probably need adjusting.